requirements.txt
anyio==4.9.0
browser-cookie3==0.20.1
certifi==2025.6.15
ffmpeg-python==0.2.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jeepney==0.9.0
lz4==4.4.4
pycryptodomex==3.23.0
python-dotenv==1.1.0
python-telegram-bot==22.1
sniffio==1.3.1
typing_extensions==4.14.0
yt-dlp==2025.6.9
